requirements.txt
# Dependências principais
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

# Banco de dados
psycopg2-binary==2.9.7
redis==4.6.0

# Processamento de dados
pandas==2.0.3
numpy==1.24.3

# Autenticação e segurança
PyJWT==2.8.0
cryptography==41.0.4

# Logging e monitoramento
structlog==23.1.0
prometheus-client==0.17.1

# Testes
pytest==7.4.2
pytest-mock==3.11.1

# Desenvolvimento
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilitários
python-dateutil==2.8.2
pytz==2023.3